+++ packages/experiment-tag/src/util/page-object.ts
@@ -1,4 +1,4 @@
-import { MessageType } from '../message-bus';
+import { MessageType } from '../subscriptions/message-bus';
 import {
   ElementAppearedTriggerValue,
   ElementVisibleTriggerValue,
